+++ src/managers/lineManager.js
@@ -1,17 +1,21 @@
-import { DOMs, DOMCtrl, DOMStrings, strings } from '../doms.js';
-import { dataStrings, saveData } from '../index.js';
+import { DOMs, DOMStrings, dataStrings, strings } from '../doms.js';
+import { saveData } from '../index.js';
 import { isValidLineName, isStartDiffersWithEnd } from '../valid.js';
+import LineUI from '../views/lineUI.js';
 
 export default class LineManager {
   constructor(stations, lines) {
     this.stations = stations;
     this.lines = lines;
+    this.UIController = new LineUI(stations, lines);
 
     this.setLineEventListeners();
   }
 
   setLineEventListeners() {
-    DOMs.LINE_MANAGER_BUTTON.addEventListener('click', this.openLineManager.bind(this));
+    DOMs.LINE_MANAGER_BUTTON.addEventListener('click', () => {
+      this.UIController.openLineManager(this.stations, this.lines);
+    });
     DOMs.MANAGER_CONTAINER.addEventListener('click', this.addLineByClick.bind(this));
     DOMs.MANAGER_CONTAINER.addEventListener('keydown', this.addLineByEnterKey.bind(this));
     DOMs.MANAGER_CONTAINER.addEventListener('click', this.deleteLineButtonClick.bind(this));
@@ -46,7 +50,7 @@ export default class LineManager {
       if (isStartDiffersWithEnd(startStation, endStation)) {
         this.lines.push(new SubwayLine(lineName, startStation, endStation));
         saveData(dataStrings.DATA_LINES, JSON.stringify(this.lines));
-        this.openLineManager.bind(this)();
+        this.UIController.openLineManager(this.stations, this.lines);
       }
     }
   }
@@ -67,70 +71,7 @@ export default class LineManager {
     const index = this.lines.findIndex(line => line.lineName === targetLine);
     this.lines.splice(index, 1);
     saveData(dataStrings.DATA_LINES, JSON.stringify(this.lines));
-    this.openLineManager.bind(this)();
-  }
-
-  openLineManager() {
-    const lineManager = `
-      <div id="${DOMStrings.LINE_MANAGER}"><br>
-        <span>${strings.LINE_NAME}</span><br>
-        <input type="text" id="${DOMStrings.LINE_NAME_INPUT}" 
-          placeholder="${strings.LINE_PLACEHOLDER}" /><br><br>
-        ${this.getLineSelectors(this.stations)}
-        <button id="${DOMStrings.LINE_ADD_BUTTON}">${strings.LINE_ADD}</button>
-        <h1>${strings.LINE_LIST_TITLE}</h1>
-        ${this.getLineList(this.lines)}
-      </div>
-    `;
-    DOMCtrl.clearManagerContainer();
-    DOMs.MANAGER_CONTAINER.innerHTML = lineManager;
-  }
-
-  getLineSelectors(stations) {
-    return `
-      <span>${strings.LINE_START} 
-        <select id="${DOMStrings.LINE_START_STATION_SELECTOR}">
-          ${stations.map(station => `<option>${station}</option>`).join('')}
-        </select>
-      </span><br>
-      <span>${strings.LINE_END} 
-        <select id="${DOMStrings.LINE_END_STATION_SELECTOR}">
-          ${stations.map(station => `<option>${station}</option>`).join('')}
-        </select>
-      </span><br><br>
-    `;
-  }
-
-  getLineList(lines) {
-    return `
-      <table id="${DOMStrings.LINE_LIST_TABLE}">
-        ${this.getLineListHeader()}
-        ${lines.map(line => this.getLineListContent(line)).join('')}
-      </table>
-    `;
-  }
-
-  getLineListHeader() {
-    return `
-      <tr>
-        <th><b>${strings.LINE_NAME}</b></th>
-        <th><b>${strings.LINE_START_STATION}</b></th>
-        <th><b>${strings.LINE_END_STATION}</b></th>
-        <th><b>${strings.SETTING}</b></th>
-      </tr>
-    `;
-  }
-
-  getLineListContent(line) {
-    return `
-      <tr>
-        <td>${line.lineName}</td>
-        <td>${line.start}</td>
-        <td>${line.end}</td>
-        <td><button class="${DOMStrings.LINE_DELETE_BUTTON}" 
-          data-${dataStrings.DATA_LINE}="${line.lineName}">${strings.DELETE}</button></td>
-      </tr>
-    `;
+    this.UIController.openLineManager(this.stations, this.lines);
   }
 }
